Moved Sound Generation to its own file
The keyboard is currently implemented in a way that prevents it from being turned into a class and easily used in other files.  To remedy this to an extent, the file that contains the keyboard code will be used as the main file and sound generation, among other features, will be done in their own files.